requirements.txt
langchain==0.3.27
langchain-openai==0.3.28
langchain-google-genai==2.1.8
langgraph==0.6.2
pandas==2.3.1
numpy==2.3.1
python-dotenv==1.1.1
pytest==8.4.1
jupyter==1.1.1
requests==2.32.4
pydantic==2.11.7
